[refactor] GoogleId class 분리
