+++ app/src/main/java/com/squirtles/musicroad/pick/DetailPickScreen.kt
@@ -57,6 +57,7 @@ import androidx.lifecycle.flowWithLifecycle
 import com.squirtles.domain.model.Pick
 import com.squirtles.musicroad.R
 import com.squirtles.musicroad.common.DialogTextButton
+import com.squirtles.musicroad.common.GoogleId
 import com.squirtles.musicroad.common.HorizontalSpacer
 import com.squirtles.musicroad.common.MessageAlertDialog
 import com.squirtles.musicroad.common.SignInAlertDialog
@@ -334,7 +335,7 @@ fun DetailPickScreen(
     if (showSignInDialogDescription != null) {
         SignInAlertDialog(
             onDismissRequest = { showSignInDialogDescription = null },
-            onGoogleSignInClick = { /*TODO*/ },
+            onGoogleSignInClick = { GoogleId(context).signIn() },
             description = showSignInDialogDescription!!
         )
     }
